+++ visualization/app/codeCharta/ui/ribbonBar/ribbonBar.component.ts
@@ -6,6 +6,7 @@ import { DownloadService } from "../../core/download/download.service";
 export class RibbonBarController {
 
     private collapsingElements = $("code-map-component #codeMap, ribbon-bar-component #header, ribbon-bar-component .section-body, #toggle-ribbon-bar-fab");
+    private toggleElements = $("ribbon-bar-component .section-title");
     private isExpanded: boolean = false;
     private _deltaMode = KindOfMap.Delta;
 
@@ -41,6 +42,14 @@ export class RibbonBarController {
     public collapse() {
         this.isExpanded = false;
         this.collapsingElements.removeClass("expanded");
+    }
+
+    public hoverToggle() {
+        this.toggleElements.addClass("toggle-hovered")
+    }
+
+    public unhoverToggle() {
+        this.toggleElements.removeClass("toggle-hovered")
     }
 
 }