+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 - Show maxValue of each metric in metricChooser select list #204
 - Colored color-slider inside the RibbonBar #318
 - Color positive buildings white and set as default option #311
+- Clicking the ribbonBar section-titles toggles the ribbonBar #324
 - Adding Linter
 
 ### Changed